fix: fixes the auto-locking issue We are using the SW persist method outlined in this SO link: https://stackoverflow.com/questions/66618136/persistent-service-worker-in-chrome-extension
As mentioned there, it is important to reconnect any connected message handlers/ports or they will be killed and the SW will restart. We were using a custom handler for a custom hook but we weren't clearing the correct function in the useEffect cleanup response, so there was a message handler still registered which got reset and the SW got killed.

This PR clears up the correct function and it seems to have fixed up the issue.